refactor: gptUtil 리팩토링
