+++ src/main/java/com/dongyang/core/controller/GptController.java
@@ -1,7 +1,7 @@
 package com.dongyang.core.controller;
 
 import com.dongyang.core.controller.dto.request.GptRequestDto;
-import com.dongyang.core.service.GptService;
+import com.dongyang.core.utils.gpt.GptUtil;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -13,12 +13,12 @@ import org.springframework.web.bind.annotation.RestController;
 @RequiredArgsConstructor
 @RequestMapping("/gpt")
 public class GptController {
-    private final GptService gptService;
+    private final GptUtil gptUtil;
 
     @GetMapping("/question")
     public ResponseEntity questionToGpt(@RequestBody GptRequestDto request) {
 
-        return ResponseEntity.ok(gptService.requestToGpt(request));
+        return ResponseEntity.ok(gptUtil.sendRequest(request));
     }
 
 }
